chore: move test data to seperate file
